Change mouse.get_(pos|rel) to output floats
In SDL3, the underlying APIs give us floats. So we should get ready to return those numbers.

+++ test/mouse_test.py
@@ -297,7 +297,7 @@ class MouseModuleTest(MouseTests):
         self.assertIsInstance(pos, tuple)
         self.assertEqual(len(pos), expected_length)
         for value in pos:
-            self.assertIsInstance(value, int)
+            self.assertIsInstance(value, float)
 
     def test_set_pos__invalid_pos(self):
         """Ensures set_pos handles invalid positions correctly."""
@@ -314,7 +314,7 @@ class MouseModuleTest(MouseTests):
         self.assertIsInstance(rel, tuple)
         self.assertEqual(len(rel), expected_length)
         for value in rel:
-            self.assertIsInstance(value, int)
+            self.assertIsInstance(value, float)
 
     def test_get_visible(self):
         """Ensures get_visible works correctly."""
